scripts/requirements.txt
openpyxl>=3.1.0
fpdf2>=2.7.0
python-pptx>=0.6.21
scipy>=1.11.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
